hexbur: hexbur provides several layers of information for Burkina Faso, particularly useful for analyzing agricultural systems

hexbur stands for "hEXagons for BURkina Faso". The basis of this package is a gridded map of Burkina Faso in hexagons of about 340 km2. For each hexagon, different information can be added using the package function. The data is not included in the package itself, but is called up via URL. It therefore requires an Internet connection to operate.

See also

Author

Maintainer: Ousmane Ouedraogo oousmane@proton.me (ORCID)